Connect to next server when backend connection fails
Client connections used to be dropped when the backend server was not
available (port closed for instance).
This commit adds a constant CONNECT_TIMEOUT_MS that specifies the amount
of time to wait for a single backend connection.
It also adds a mechanism to try the next server when timeout is reached
or an error happens.